biome_lib.register_on_generate: Adding `density` field which provies a different method of specifying the density of the spawning.
If set, `max_count` and `rarity` are ignored.

`density` is simply a percentage of the available nodes that spawning will be attempted on. If this results in a number of nodes that isn't a whole number, e.g. 3.7, then this will be randomly either rounded up or down. The chance of rounding up is proportional to how close it is to the ceiling, so for 3.7 it will be rounded up to 4 70% of the time. This allows very low densities to be specified that will actually be accurate, and not inaccurate due to rounding.

Also: When we randomly pick a node from the available nodes, that node is removed from the list. This stops the same node being randomly selected twice, improving accuracy without any measureable performance impact.